in progress: promising traces in exceptions
