+++ examples/agents_with_errors.py
@@ -1,50 +1,4 @@
-import sys
-import traceback
-
 import promising
-
-
-# TODO [TRACES] Move this hook somewhere in the promising framework itself
-def my_excepthook(exc_type, exc_value, exc_tb):
-    # TODO [TRACES] Fallback to default printing behavior if the exception does not have
-    #  __promising_context__  attribute at all
-    # TODO [TRACES] Is it possible to fetch the width of the terminal and use it for the
-    #  horizontal line length ?
-    print("━" * 60)
-    print(f"💥  {exc_type.__name__}: {exc_value}")
-    print("━" * 60)
-    traceback.print_tb(exc_tb)
-    print("━" * 60)
-
-    pc = getattr(exc_value, "__promising_context__", None)
-    if pc is None:
-        return
-
-    print("📍  Promise creation stacks (outermost → innermost):")
-    print("━" * 60)
-    for ctx in pc.get_trace(parents_first=True):
-        if not isinstance(ctx, promising.Promise):
-            continue
-        stack_summary = getattr(ctx, "_creation_stack_summary", None)
-        if stack_summary is None:
-            continue
-        print(f"{ctx!r}")
-        for line in stack_summary.format():
-            print(line, end="")
-        print("━" * 60)
-    # TODO [TRACES] At the very end the final traceback should be printed in
-    #  the same filtered fashion - the actual traceback of the exception that
-    #  was raised
-    # TODO [TRACES] Make sure something like this is printed everytime
-    #  framework frames are omitted ?
-    #  `... (`promising` internals omitted) ...`
-    # TODO [TRACES] Do the same with `asyncio` and simplify skipping logic
-    #  (process whole trace - don't stop at framework frames)
-
-
-# TODO [TRACES] What about formatting it for the loggers, and not just
-#  stderr/stdout ?
-sys.excepthook = my_excepthook
 
 
 def agent3_plain_func2() -> None:
